--- src/components/CharacterList.tsx
@@ -1,722 +1,0 @@
-import { IChar, IState } from '../types/Types';
-
-const imageFolder = window.location.href.includes('local') ? `${window.location.href}images/character_icons/` : 'https://wischli.github.io/smashbros-random-with-bans/images/character_icons';
-// const imageFolder = `https://wischli.github.io/smashbros-random-with-bans/images/character_icons`;
-function imigify(name: string) {
-  const title = name
-    .split(/[. \-&]/g)
-    .map(item => item.toLowerCase())
-    .join('');
-  return `${imageFolder}/${title}.png`;
-}
-
-const CharacterList: IChar[] = [
-  {
-    id: 1,
-    name: 'Mario',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 43,
-    name: 'Dr. Mario',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 2,
-    name: 'Luigi',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 3,
-    name: 'Peach',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [3.5],
-    display: true,
-  },
-  {
-    id: 3.5,
-    name: 'Daisy',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [3],
-    display: true,
-  },
-  {
-    id: 4,
-    name: 'Bowser',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 5,
-    name: 'Yoshi',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 6,
-    name: 'Rosalina & Luma',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 7,
-    name: 'Donkey Kong',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 8,
-    name: 'Diddy Kong',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 9,
-    name: 'Link',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 10,
-    name: 'Zelda',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 11,
-    name: 'Sheik',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 45,
-    name: 'Ganondorf',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 12,
-    name: 'Toon Link',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 12.9,
-    name: 'Young Link',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 13,
-    name: 'Samus',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [13.5],
-    display: true,
-  },
-  {
-    id: 13.5,
-    name: 'Dark Samus',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [13],
-    display: true,
-  },
-  {
-    id: 14,
-    name: 'Zero Suit Samus',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 15,
-    name: 'Pit',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [15],
-    display: true,
-  },
-  {
-    id: 15.5,
-    name: 'Dark Pit',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [15.5],
-    display: true,
-  },
-  {
-    id: 16,
-    name: 'Palutena',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 17,
-    name: 'Marth',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [17.5],
-    display: true,
-  },
-  {
-    id: 17.5,
-    name: 'Lucina',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [17],
-    display: true,
-  },
-  {
-    id: 18,
-    name: 'Ike',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 19,
-    name: 'Robin',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 20,
-    name: 'Kirby',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 21,
-    name: 'King Dedede',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 22,
-    name: 'Meta Knight',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 23,
-    name: 'Little Mac',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 24,
-    name: 'Fox',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 25,
-    name: 'Pikachu',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 25.9,
-    name: 'Pichu',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 26,
-    name: 'Pokemon Trainer',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  // {
-  //   id: 26,
-  //   name: 'Charizard',
-  //   media: '',
-  //   nicknames: [],
-  //   enabled: true,
-  //   played: false,
-  //   randomOrder: -1,
-  //   echo: [],
-  //   display: true,
-  //   moves: []
-  // },
-  // {
-  //   id: 26.7,
-  //   name: 'Ivysaur',
-  //   media: '',
-  //   nicknames: [],
-  //   enabled: true,
-  //   played: false,
-  //   randomOrder: -1,
-  //   echo: [],
-  //   display: true,
-  //   moves: []
-  // },
-  // {
-  //   id: 26.8,
-  //   name: 'Squirtle',
-  //   media: '',
-  //   nicknames: [],
-  //   enabled: true,
-  //   played: false,
-  //   randomOrder: -1,
-  //   echo: [],
-  //   display: true,
-  //   moves: []
-  // },
-  {
-    id: 27,
-    name: 'Lucario',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 28,
-    name: 'Greninja',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 29,
-    name: 'Captain Falcon',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 30,
-    name: 'Villager',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 31,
-    name: 'Olimar',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 32,
-    name: 'Wii Fit Trainer',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 33,
-    name: 'Shulk',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 34,
-    name: 'Pac-Man',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 35,
-    name: 'Mega Man',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 36,
-    name: 'Sonic',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 38,
-    name: 'Ness',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 39,
-    name: 'Falco',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 40,
-    name: 'Wario',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 41,
-    name: 'Snake',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 42,
-    name: 'Ice Climbers',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 44,
-    name: 'R.O.B.',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 46,
-    name: 'Mr. Game & Watch',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 47,
-    name: 'Bowser Jr.',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 48,
-    name: 'Duck Hunt',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 49,
-    name: 'Jigglypuff',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 50,
-    name: 'Mewtwo',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 51,
-    name: 'Lucas',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 52,
-    name: 'Roy',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [52.5],
-    display: true,
-  },
-  {
-    id: 52.5,
-    name: 'Chrom',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [52],
-    display: true,
-  },
-  {
-    id: 53,
-    name: 'Ryu',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [53.5],
-    display: true,
-  },
-  {
-    id: 53.5,
-    name: 'Ken',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [53],
-    display: true,
-  },
-  {
-    id: 54,
-    name: 'Cloud',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 55,
-    name: 'Corrin',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 56,
-    name: 'Bayonetta',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 57,
-    name: 'Mii Fighter Brawler',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 58,
-    name: 'Mii Fighter Shooter',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 59,
-    name: 'Mii Fighter Sword',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 60,
-    name: 'Inkling',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 61,
-    name: 'Ridley',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 62,
-    name: 'Simon',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [62.5],
-    display: true,
-  },
-  {
-    id: 62.5,
-    name: 'Richter',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [62],
-    display: true,
-  },
-  {
-    id: 63,
-    name: 'King K. Rool',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 64,
-    name: 'Isabelle',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 65,
-    name: 'Incineroar',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-  {
-    id: 66,
-    name: 'Piranha Plant',
-    media: '',
-    enabled: true,
-    played: false,
-    echo: [],
-    display: true,
-  },
-].map((char: IChar) => {
-  return {...char, media: imigify(char.name)};
-});
-// const initialCharState: IState = { enabled: CharacterList, played: [], disabled: [], hidden: [] };
-const initialCharState: IState = { enabled: CharacterList.map((char, index) => index), played: [], disabled: [], hidden: [] };
-
-export { CharacterList, initialCharState, imigify };
