+++ src/components/CookieNotice/CookieNotice-container.tsx
@@ -1,9 +1,10 @@
 import React, { useContext, useState } from 'react';
 import Cookies from 'universal-cookie';
-import Context, { Icontext } from './Context';
+import Context, { Icontext } from '../Context';
+import { buttonStyle, messageStyle, noticeStyle } from './CookieNotice-style';
 
 const CookieNotice = () => {
-  const { themeStyle, cookies } = useContext(Context as React.Context<Icontext>);
+  const { cookies } = useContext(Context as React.Context<Icontext>);
   const [hideNotice, setHide] = useState(cookies.notice);
 
   const handleCloseClick = () => {
@@ -11,27 +12,10 @@ const CookieNotice = () => {
     cookyInstance.set('notice', JSON.stringify(true), { path: '/' });
     setHide(true);
   };
-  const styles = {
-    cookieNotice: {
-      bottom: 0,
-      backgroundColor: themeStyle.bgCookieNotice,
-      color: 'white',
-      width: '100%',
-      display: cookies.notice ? 'block' : 'none',
-      zIndex: 1,
-    },
-    cookieMessage: {
-      padding: 10,
-      display: hideNotice ? 'none' : 'flex',
-      alignItems: 'center',
-      maxWidth: '100%',
-      justifyContent: window.innerWidth < 800 ? 'center' : 'space-evenly',
-    },
-  };
 
   const CookieMessage = () => {
     return (
-      <div style={styles.cookieMessage}>
+      <div style={messageStyle(!hideNotice)}>
         <div>
           This website uses cookies to store your current state of characters. For more information, see
           <a style={{ color: '#9eef8b' }} href="https://github.com/wischli/smashbros-random-with-bans">
@@ -39,7 +23,7 @@ const CookieNotice = () => {
           </a>
           .
         </div>
-        <button type="button" style={{ ...themeStyle.button, minWidth: 100 }} onClick={() => handleCloseClick()}>
+        <button type="button" style={ buttonStyle } onClick={() => handleCloseClick()}>
           Accept
         </button>
       </div>
@@ -47,7 +31,7 @@ const CookieNotice = () => {
   };
 
   return (
-    <div style={{ ...styles.cookieNotice, position: 'fixed' }}>
+    <div style={noticeStyle(cookies.notice)}>
       <CookieMessage />
     </div>
   );
