+++ src/index.tsx
@@ -1,7 +1,7 @@
 /* global document */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import './layout/index.css';
+import App from './app/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
