+++ src/layout/themeStyle.ts
@@ -1,0 +1,29 @@
+export const color = {
+  background: '#486471',
+  bgContent: '#486471',
+  red: '#ffb3b3',
+  green: '#99f9ae',
+  card: '#26353c',
+  cardTitle: '#aacdde',
+  nav: '#466370',
+  bgNavSecBtn: '#ddddddb3',
+  bgCookieNotice: '#26353c',
+  charEnabled: 'white',
+  charDisabled: '#ff6767b8',
+  charPlayed: '#70f78e',
+  charHidden: 'grey',
+  cardBtnLeft: '#383838',
+  cardBtnRight: 'black',
+};
+
+export const themeButtonStyle: React.CSSProperties = {
+  height: '50px',
+  fontSize: '15px',
+  fontWeight: 700,
+  textAlign: 'center',
+  justifyContent: 'center',
+  borderRadius: '5px',
+  textTransform: 'uppercase',
+  letterSpacing: '1px',
+  border: '2px solid',
+};
