+++ src/model/charArray/characters.ts
@@ -1,0 +1,376 @@
+export const characters: {
+  id: number,
+  name: string,
+  echo: number[],
+}[] = [
+    {
+    id: 1,
+    name: 'Mario',
+    echo: [],
+  },
+  {
+    id: 43,
+    name: 'Dr. Mario',
+    echo: [],
+  },
+  {
+    id: 2,
+    name: 'Luigi',
+    echo: [],
+  },
+  {
+    id: 3,
+    name: 'Peach',
+    echo: [3.5],
+  },
+  {
+    id: 3.5,
+    name: 'Daisy',
+    echo: [3],
+  },
+  {
+    id: 4,
+    name: 'Bowser',
+    echo: [],
+  },
+  {
+    id: 5,
+    name: 'Yoshi',
+    echo: [],
+  },
+  {
+    id: 6,
+    name: 'Rosalina & Luma',
+    echo: [],
+  },
+  {
+    id: 7,
+    name: 'Donkey Kong',
+    echo: [],
+  },
+  {
+    id: 8,
+    name: 'Diddy Kong',
+    echo: [],
+  },
+  {
+    id: 9,
+    name: 'Link',
+    echo: [],
+  },
+  {
+    id: 10,
+    name: 'Zelda',
+    echo: [],
+  },
+  {
+    id: 11,
+    name: 'Sheik',
+    echo: [],
+  },
+  {
+    id: 45,
+    name: 'Ganondorf',
+    echo: [],
+  },
+  {
+    id: 12,
+    name: 'Toon Link',
+    echo: [],
+  },
+  {
+    id: 12.9,
+    name: 'Young Link',
+    echo: [],
+  },
+  {
+    id: 13,
+    name: 'Samus',
+    echo: [13.5],
+  },
+  {
+    id: 13.5,
+    name: 'Dark Samus',
+    echo: [13],
+  },
+  {
+    id: 14,
+    name: 'Zero Suit Samus',
+    echo: [],
+  },
+  {
+    id: 15,
+    name: 'Pit',
+    echo: [15],
+  },
+  {
+    id: 15.5,
+    name: 'Dark Pit',
+    echo: [15.5],
+  },
+  {
+    id: 16,
+    name: 'Palutena',
+    echo: [],
+  },
+  {
+    id: 17,
+    name: 'Marth',
+    echo: [17.5],
+  },
+  {
+    id: 17.5,
+    name: 'Lucina',
+    echo: [17],
+  },
+  {
+    id: 18,
+    name: 'Ike',
+    echo: [],
+  },
+  {
+    id: 19,
+    name: 'Robin',
+    echo: [],
+  },
+  {
+    id: 20,
+    name: 'Kirby',
+    echo: [],
+  },
+  {
+    id: 21,
+    name: 'King Dedede',
+    echo: [],
+  },
+  {
+    id: 22,
+    name: 'Meta Knight',
+    echo: [],
+  },
+  {
+    id: 23,
+    name: 'Little Mac',
+    echo: [],
+  },
+  {
+    id: 24,
+    name: 'Fox',
+    echo: [],
+  },
+  {
+    id: 25,
+    name: 'Pikachu',
+    echo: [],
+  },
+  {
+    id: 25.9,
+    name: 'Pichu',
+    echo: [],
+  },
+  {
+    id: 26,
+    name: 'Pokemon Trainer',
+    echo: [],
+  },
+  {
+    id: 27,
+    name: 'Lucario',
+    echo: [],
+  },
+  {
+    id: 28,
+    name: 'Greninja',
+    echo: [],
+  },
+  {
+    id: 29,
+    name: 'Captain Falcon',
+    echo: [],
+  },
+  {
+    id: 30,
+    name: 'Villager',
+    echo: [],
+  },
+  {
+    id: 31,
+    name: 'Olimar',
+    echo: [],
+  },
+  {
+    id: 32,
+    name: 'Wii Fit Trainer',
+    echo: [],
+  },
+  {
+    id: 33,
+    name: 'Shulk',
+    echo: [],
+  },
+  {
+    id: 34,
+    name: 'Pac-Man',
+    echo: [],
+  },
+  {
+    id: 35,
+    name: 'Mega Man',
+    echo: [],
+  },
+  {
+    id: 36,
+    name: 'Sonic',
+    echo: [],
+  },
+  {
+    id: 38,
+    name: 'Ness',
+    echo: [],
+  },
+  {
+    id: 39,
+    name: 'Falco',
+    echo: [],
+  },
+  {
+    id: 40,
+    name: 'Wario',
+    echo: [],
+  },
+  {
+    id: 41,
+    name: 'Snake',
+    echo: [],
+  },
+  {
+    id: 42,
+    name: 'Ice Climbers',
+    echo: [],
+  },
+  {
+    id: 44,
+    name: 'R.O.B.',
+    echo: [],
+  },
+  {
+    id: 46,
+    name: 'Mr. Game & Watch',
+    echo: [],
+  },
+  {
+    id: 47,
+    name: 'Bowser Jr.',
+    echo: [],
+  },
+  {
+    id: 48,
+    name: 'Duck Hunt',
+    echo: [],
+  },
+  {
+    id: 49,
+    name: 'Jigglypuff',
+    echo: [],
+  },
+  {
+    id: 50,
+    name: 'Mewtwo',
+    echo: [],
+  },
+  {
+    id: 51,
+    name: 'Lucas',
+    echo: [],
+  },
+  {
+    id: 52,
+    name: 'Roy',
+    echo: [52.5],
+  },
+  {
+    id: 52.5,
+    name: 'Chrom',
+    echo: [52],
+  },
+  {
+    id: 53,
+    name: 'Ryu',
+    echo: [53.5],
+  },
+  {
+    id: 53.5,
+    name: 'Ken',
+    echo: [53],
+  },
+  {
+    id: 54,
+    name: 'Cloud',
+    echo: [],
+  },
+  {
+    id: 55,
+    name: 'Corrin',
+    echo: [],
+  },
+  {
+    id: 56,
+    name: 'Bayonetta',
+    echo: [],
+  },
+  {
+    id: 57,
+    name: 'Mii Fighter Brawler',
+    echo: [],
+  },
+  {
+    id: 58,
+    name: 'Mii Fighter Shooter',
+    echo: [],
+  },
+  {
+    id: 59,
+    name: 'Mii Fighter Sword',
+    echo: [],
+  },
+  {
+    id: 60,
+    name: 'Inkling',
+    echo: [],
+  },
+  {
+    id: 61,
+    name: 'Ridley',
+    echo: [],
+  },
+  {
+    id: 62,
+    name: 'Simon',
+    echo: [62.5],
+  },
+  {
+    id: 62.5,
+    name: 'Richter',
+    echo: [62],
+  },
+  {
+    id: 63,
+    name: 'King K. Rool',
+    echo: [],
+  },
+  {
+    id: 64,
+    name: 'Isabelle',
+    echo: [],
+  },
+  {
+    id: 65,
+    name: 'Incineroar',
+    echo: [],
+  },
+  {
+    id: 66,
+    name: 'Piranha Plant',
+    echo: [],
+  },
+];
